ODE-310 Renamed for naming consistency with VSD and BSM features.
